+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+-   Cat command.
+
 ## [1.9.0-beta 2023-11-23]
 
 ### Added