+++ CHANGELOG.md
@@ -7,6 +7,12 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 -   Cat command.
+-   Cowsay command.
+-   TextHelper.ConcatenateSentence method.
+
+### Changed
+
+-   The terminal now distinguishes between sentences wrapped in " or ' and treats them as a whole.
 
 ## [1.9.0-beta 2023-11-23]